+++ R/load_data.R
@@ -2,6 +2,7 @@
 #' @param ticker ticker code from Yahoo.
 #' @param src source name ("yahoo" or "FRED")
 #' @param from load data `from` untill today where the format of `from` is yyyy-mm-dd.
+#' @param prefix logical (default is TRUE) indicating if ticker name should be included in column name of columns Change and ChangeSinceStart (only applicable for FRED data)
 #' @return A data.frame
 #' @examples
 #' # Load NOVO data since 2021-01-01
@@ -13,7 +14,9 @@
 #' data
 #' @export
 #'
-load_data <- function(ticker, src = "yahoo", from = "2014-01-01") {
+load_data <- function(ticker, src = "yahoo", from = "2014-01-01", prefix = TRUE) {
+
+  options("getSymbols.warning4.0" = FALSE)
 
   data <- catch_error(
     quantmod::getSymbols(ticker, auto.assign = FALSE, from = from, src = src)
@@ -33,16 +36,17 @@ load_data <- function(ticker, src = "yahoo", from = "2014-01-01") {
     data <- data %>%
       dplyr::rename_with(.data = ., ~ gsub(paste0(ticker, "."), "", .x)) %>%
       dplyr::mutate(Change = price_change(Close, digits = 4),
-                    ChangeFromStart = prince_change_from_start(Close, digits = 4)) %>%
-      dplyr::relocate(Date, Open, High, Low, Close, Adjusted, Change, ChangeFromStart, Volume)
+                    ChangeSinceStart = prince_change_since_start(Close, digits = 4)) %>%
+      dplyr::relocate(Date, Open, High, Low, Close, Adjusted, Change, ChangeSinceStart, Volume)
 
   } else {
 
     data <- data %>%
       dplyr::rename_with(~ gsub(ticker, "Close", .x)) %>%
       dplyr::mutate(Change = price_change(Close, digits = 4),
-                    ChangeFromStart = prince_change_from_start(Close, digits = 4)) %>%
-      dplyr::relocate(Date, Close, Change, ChangeFromStart)
+                    ChangeSinceStart = prince_change_since_start(Close, digits = 4)) %>%
+      dplyr::relocate(Date, Close, Change, ChangeSinceStart) %>%
+      rename_cols(ticker = ticker, rename = prefix)
 
   }
 
@@ -52,10 +56,41 @@ load_data <- function(ticker, src = "yahoo", from = "2014-01-01") {
 
 #' Change from start
 #' @param x numeric vector
+#' @param start integer specifying start element (default is 1, i.e. the first element of x)
 #' @param digits number of digits using in rounding
+#' @return a numeric vector of same length as x
+#' @export
 #'
-prince_change_from_start <- function(x, digits = 4) {
+prince_change_since_start <- function(x, start = 1, digits = 4) {
 
-  round(100 * ((x - x[1]) / x), digits = digits)
+  round(100 * ((x - x[start]) / x), digits = digits)
+
+}
+
+#' Rename columns in data.frame
+#' @param data data.frame with columns Close, Change and ChangeSinceStart
+#' @param ticker character string with ticker name
+#' @param rename logical (default is TRUE). If FALSE, data is returned without any renaming
+#' @return data.frame
+#' @description Renaming columns from:
+#' * Close -> ticker
+#' * Change -> ticker_Change
+#' * ChangeSinceStart -> ticker_ChangeSinceStart
+#' @details This is a helper function for `load_data`
+rename_cols <- function(data, ticker, rename = TRUE) {
+
+  if (!rename) {
+    return(data)
+  }
+
+  str_change <- paste0(ticker, "_Change")
+  str_ChangeSinceStart <- paste0(ticker, "_ChangeSinceStart")
+
+  data %>%
+    dplyr::rename(
+      !!ticker := Close,
+      !!str_change := Change,
+      !!str_ChangeSinceStart := ChangeSinceStart
+    )
 
 }