+++ tests/testthat/test-load_data.R
@@ -2,7 +2,7 @@ test_that("Load yahoo data works", {
 
   data <- load_data(ticker = "NOVO-B.CO", from = "2021-01-01")
 
-  cols <- c("Date", "Open", "High", "Low", "Close", "Adjusted", "Change", "ChangeFromStart", "Volume")
+  cols <- c("Date", "Open", "High", "Low", "Close", "Adjusted", "Change", "ChangeSinceStart", "Volume")
 
   ## Expecting identical column names
   expect_identical(names(data), cols)
@@ -27,17 +27,24 @@ test_that("Load FRED data works", {
 
   data <- load_data(ticker = "MORTGAGE30US", src = "FRED")
 
-  cols <- c("Date", "Close", "Change", "ChangeFromStart")
+  cols <- c("Date", "MORTGAGE30US", "MORTGAGE30US_Change", "MORTGAGE30US_ChangeSinceStart")
 
   ## Expecting identical column names
   expect_identical(names(data), cols)
 
   price <- data %>%
     dplyr::filter(Date == "2022-01-06") %>%
-    dplyr::pull(Close) %>%
+    dplyr::pull(MORTGAGE30US) %>%
     round(2)
 
   ## Expecting a certain closing value in the first row
   expect_equal(price, 3.22)
+
+  data2 <- load_data(ticker = "MORTGAGE30US", src = "FRED", prefix = FALSE)
+
+  cols <- c("Date", "Close", "Change", "ChangeSinceStart")
+
+  ## Expecting identical column names
+  expect_identical(names(data2), cols)
 
 })