+++ NEWS.md
@@ -1,3 +1,8 @@
+# kb.yahoo 1.5.0
+
+* Adding `prefix` parameter to `load_data`
+* Removing start up message from `quantmod`
+
 # kb.yahoo 1.4.0
 
 * Adding column ChangeFromStart to output of `load_data`